refactor(media): move VlcRuntime to mur-common (shared with runtime)
VlcRuntime + runtime_path were defined in mur-core/src/cmd/media/mod.rs
but are needed by mur-agent-runtime (WatchScheduler, Plan B). Because
core → runtime, not the reverse, the plain serde types are moved down
to mur-common which both crates depend on. mur-core re-exports them
unchanged so all call-sites continue to compile without modification.

Removed the now-unused `serde` and `PathBuf` imports from mod.rs.
