+++ mur-common/src/lib.rs
@@ -27,6 +27,7 @@ pub mod llm;
 pub mod local_llm;
 pub mod lock_file;
 pub mod manifest;
+pub mod media;
 pub mod mobile;
 pub mod model;
 pub mod model_resolve;
